:white_check_mark: add two e2e tests
